TP-897: Moved GigaspaceMetricsExporter to GsModule, rather than a separate GsMetricsModule.
