+++ astrix-gs/src/main/java/com/avanza/astrix/gs/localview/GsLocalViewComponent.java
@@ -39,8 +39,8 @@ import com.avanza.astrix.config.DynamicLongProperty;
 import com.avanza.astrix.core.util.ReflectionUtil;
 import com.avanza.astrix.gs.ClusteredProxyBinder;
 import com.avanza.astrix.gs.GigaSpaceProxy;
+import com.avanza.astrix.gs.GigaspaceMetricsExporter;
 import com.avanza.astrix.gs.GsBinder;
-import com.avanza.astrix.gs.metrics.GigaspaceMetricsExporter;
 import com.avanza.astrix.provider.component.AstrixServiceComponentNames;
 import com.avanza.astrix.spring.AstrixSpringContext;
 import com.j_spaces.core.IJSpace;
