+++ astrix-integration-tests/src/test/java/com/avanza/astrix/integration/tests/GigaSpacesAuthenticationTest.java
@@ -33,8 +33,8 @@ import com.avanza.astrix.beans.tracing.DefaultTraceProvider;
 import com.avanza.astrix.config.GlobalConfigSourceRegistry;
 import com.avanza.astrix.config.MapConfigSource;
 import com.avanza.astrix.gs.ClusteredProxyCacheImpl;
+import com.avanza.astrix.gs.GigaspaceMetricsExporter;
 import com.avanza.astrix.gs.GsBinder;
-import com.avanza.astrix.gs.metrics.GigaspaceMetricsExporter;
 import com.avanza.astrix.gs.security.DefaultGsSecurityProvider;
 import com.avanza.astrix.gs.security.GsSecurityProvider;
 import com.avanza.gs.test.PuConfigurers;
@@ -78,10 +78,9 @@ public class GigaSpacesAuthenticationTest {
 			return new DefaultCredentialsProvider(clientCredentials);
 		}
 	};
-	private final GigaspaceMetricsExporter metricsExporter = GigaspaceMetricsExporter.noExporter();
 	private final ServiceProperties lunchPuServiceProperties = new GsBinder().createProperties(LUNCH_PU.getClusteredGigaSpace());
 	private final ServiceProperties lunchGraderPuServiceProperties = new GsBinder().createProperties(LUNCH_GRADER_PU.getClusteredGigaSpace());
-	private final ClusteredProxyCacheImpl proxyCache = new ClusteredProxyCacheImpl(traceProvider, mockGsSecurityProvider, metricsExporter);
+	private final ClusteredProxyCacheImpl proxyCache = new ClusteredProxyCacheImpl(traceProvider, mockGsSecurityProvider, GigaspaceMetricsExporter.noExporter());
 
 	/**
 	 * This are the client-side credentials that will be used when connecting to
